+++ app.js
@@ -31,6 +31,7 @@ app.use(function(req, res, next) {
 // Controllers
 app.use('/auth', require('./routes/auth'));
 app.use('/calendar', require('./routes/calendar'));
+app.use('/calendar/event', require('./routes/event'));
 
 app.get('*', function(req, res, next) {
 	res.sendFile(path.resolve(__dirname, 'client', 'build', 'index.html'));
